% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{DensityDPOne}
\alias{DensityDPOne}
\title{{DensityDPOne}}
\usage{
DensityDPOne(para)
}
\arguments{
\item{para}{parameter list.}
}
\value{
An updated parameter list.
}
\description{
Given regulation type (z value). Update density specification based on DPM fitting.
}
\keyword{internal}
