% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{DPdensCluster}
\alias{DPdensCluster}
\title{{DPdensCluster}}
\usage{
DPdensCluster(para, twogrps = FALSE)
}
\arguments{
\item{para}{The parameter list}
}
\value{
The updated parameter list
}
\description{
First, apply DP-package to split data into several small normal groups based on their test statistics, then merge until we attain total three groups based on K-L distance between proposed null density vs. prior null density from biological background knowledge. Finally, return each group's mixture normal parameters as a list.
}
\keyword{internal}
