% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{DPdensitySubset}
\alias{DPdensitySubset}
\title{{DPdensitySubset}}
\usage{
DPdensitySubset(ll, subdat, subprior, submcmc, substatus)
}
\arguments{
\item{ll}{NULL}

\item{subdat}{NULL}

\item{subprior}{NULL}

\item{submcmc}{NULL}

\item{substatus}{NULL}
}
\value{
para
}
\description{
DPdensitySubset
}
\keyword{internal}
