% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{GCut.z}
\alias{GCut.z}
\title{{GCut.z}}
\usage{
GCut.z(para)
}
\arguments{
\item{para}{The parameter list}
}
\value{
The updated parameter list
}
\description{
Split network into three where nodes located within each group has the same class label (z value).
}
\keyword{internal}
