% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{hyperParaDMH}
\alias{hyperParaDMH}
\title{{hyperParaDMH}}
\usage{
hyperParaDMH(para, plot = TRUE)
}
\arguments{
\item{para}{The parameter list}

\item{plot}{Logical. T: plot trace plot; F: not plot output.}
}
\value{
a list:
\item{rhoTrack}{saved rho value each iteration}
\item{piTrack}{saved pi value each iteration}
\item{zTrack}{saved z label value each iteration}
\item{rho}{posterior mean calculated based on second half of the iterations}
\item{zstat}{posterior z value calculated based on major vote method.}
}
\description{
Apply Double Metropolis-Hastings sampler (DMH) to find hyperparameters: rho & pi
}
\keyword{internal}
