% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{llikDMH}
\alias{llikDMH}
\title{{llikDMH}}
\usage{
llikDMH(para, z, rho, pi)
}
\arguments{
\item{para}{NULL}

\item{z}{NULL}

\item{rho}{NULL}

\item{pi}{NULL}
}
\value{
llik Value
}
\description{
llikDMH
}
\keyword{internal}
