% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{logdensMixNorm}
\alias{logdensMixNorm}
\title{{logdensMixNorm}}
\usage{
logdensMixNorm(dat, mu, sd, prop)
}
\arguments{
\item{dat}{NULL}

\item{mu}{NULL}

\item{sd}{NULL}

\item{prop}{NULL}
}
\value{
log density value
}
\description{
logdensMixNorm
}
\keyword{internal}
