% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missloc.R
\docType{data}
\name{missloc}
\alias{missloc}
\title{Gene missing location dataset}
\format{A vector of length=157, containing the missing location, resulting in a 20\% missingness in the simulation study.}
\usage{
data(missloc)
}
\description{
The simulated missing locations
}
\details{
The missing location are randomly selected among 'non-hub' genes for the simulated network with 776 gene nodes.
}
\keyword{datasets}
