% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\docType{data}
\name{net}
\alias{net}
\title{The network used in simulation studies}
\format{A 0/1 adjacency matrix.}
\usage{
data(net)
}
\description{
The network used in simulation studies extracted from real data network, represented by adjacency matrix 0/1.
}
\details{
In the simulation studies, the network is a subnetwork extracted based on real network, used in simulation studies. It contains 776 gene nodes.
}
\keyword{datasets}
