% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{r.BayesImpute}
\alias{r.BayesImpute}
\title{{r.BayesImpute}}
\usage{
r.BayesImpute(para)
}
\arguments{
\item{para}{parameter list.}
}
\value{
An updated parameter list.
}
\description{
Impute missing node by sampling from posterior density.
}
\keyword{internal}
