% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{r.knnImpute}
\alias{r.knnImpute}
\title{{r.knnImpute}}
\usage{
r.knnImpute(para)
}
\arguments{
\item{para}{parameter list.}
}
\value{
An updated parameter list.
}
\description{
Impute missing node using their nearest neighbor test statistics.
}
\keyword{internal}
