% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{SWCut}
\alias{SWCut}
\title{{SWCut}}
\usage{
SWCut(para)
}
\arguments{
\item{para}{parameter list}

\item{rhoScale}{a 4-element vector, used when rho values need to be scaled for a propor graph cut. Default=c(1,1,1,0)}
}
\value{
A list with three elements, representing z=1, z=2, or z=3. Each element is also a list, contains all subgraph with same z value.
}
\description{
Swendsen-Wang graph cut techniques. The edges located within each subgraph (same z) has a positive probability to stay "on" or "off". The edges with "off" labels are closed so that it is further cut to smaller subgraphs.
}
\keyword{internal}
