% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{z.update.fastSW}
\alias{z.update.fastSW}
\title{{z.update.fastSW}}
\usage{
z.update.fastSW(para, swcutList)
}
\arguments{
\item{para}{The parameter list}

\item{swcutList}{a list, the direct output from SWCut.}
}
\value{
An updated parameter list.
}
\description{
Given density specification. Update regulation type (z value) within each subgraph cut by Swendsen-Wang graph cut techniques.
}
\keyword{internal}
