% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslabel.R
\name{z.update.Mclustprior}
\alias{z.update.Mclustprior}
\title{{z.update.Mclustprior}}
\usage{
z.update.Mclustprior(para, oldz, rho, pi)
}
\arguments{
\item{para}{The parameter list}

\item{oldz}{NULL}

\item{rho}{NULL}

\item{pi}{NULL}
}
\value{
voted z
}
\description{
z.update.Mclustprior
}
\keyword{internal}
